Move authorization logic to spotify api client
